requirements.txt
playwright
requests
google-cloud-storage